requirements.txt
bcrypt==3.1.0
cffi==1.7.0
click==6.6
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
pycparser==2.14
python-ldap==2.4.25
six==1.10.0
SQLAlchemy==1.0.14
Werkzeug==0.11.10
wheel==0.24.0
